Frontend/requirements.txt
django~=2.2
djangorestframework
psycopg2
pyjwt
requests
django-cors-headers
